increase channel size

+++ proxy/shmring.h
@@ -18,8 +18,8 @@ struct ring_buffer {
 
 struct ring_buffer* shmring_init(void *base_addr, size_t size);
 void shmring_reset(struct ring_buffer *ring, size_t size);
-int shmring_pop(struct ring_buffer *rx_ring, void *buf, size_t size);
-int shmring_read(struct ring_buffer *rx_ring, void *buf, size_t size);
-int shmring_push(struct ring_buffer *tx_ring, void *buf, size_t size);
+size_t shmring_pop(struct ring_buffer *rx_ring, void *buf, size_t size);
+size_t shmring_read(struct ring_buffer *rx_ring, void *buf, size_t size);
+size_t shmring_push(struct ring_buffer *tx_ring, void *buf, size_t size);
 
 #endif /* ndef SHMRING_H_ */
